devices/analog-meter-readers/drum-meter/requirements.txt
paddleocr==2.3.0.1
paddlepaddle==2.1.3
paho-mqtt==1.5.1
python-dotenv==0.19.0
opencv-python==4.5.3.56
